refactor(combos): split create/edit into shared form and migrate partials
Extracts a shared useComboForm hook and ComboForm used by create and edit, a
useAddProduct hook with a reusable VariantChecklist, and a ComboProductItem
row, then moves everything out of the legacy partials/ folder into
components/. Unifies the min=0 guard on both number inputs. Refs #81.
